+++ sqlsubscribers/margin_levels.go
@@ -1,6 +1,7 @@
 package sqlsubscribers
 
 import (
+	"context"
 	"time"
 
 	"code.vegaprotocol.io/data-node/entities"
@@ -17,13 +18,15 @@ type MarginLevelsEvent interface {
 
 //go:generate go run github.com/golang/mock/mockgen -destination mocks/margin_levels_mock.go -package mocks code.vegaprotocol.io/data-node/sqlsubscribers MarginLevelsStore
 type MarginLevelsStore interface {
-	Upsert(*entities.MarginLevels) error
+	Add(*entities.MarginLevels) error
+	OnTimeUpdateEvent(context.Context) error
 }
 
 type MarginLevels struct {
 	store    MarginLevelsStore
 	log      *logging.Logger
 	vegaTime time.Time
+	seqNum   uint64
 }
 
 func NewMarginLevels(store MarginLevelsStore, log *logging.Logger) *MarginLevels {
@@ -41,8 +44,13 @@ func (ml *MarginLevels) Push(evt events.Event) error {
 	switch e := evt.(type) {
 	case TimeUpdateEvent:
 		ml.vegaTime = e.Time()
+		err := ml.store.OnTimeUpdateEvent(e.Context())
+		if err != nil {
+			ml.log.Error("inserting margin level events to Postgres failed", logging.Error(err))
+		}
 	case MarginLevelsEvent:
-		return ml.consume(e)
+    ml.seqNum = e.Sequence()
+		ml.consume(e)
 	default:
 		return errors.Errorf("unknown event type %s", e.Type().String())
 	}
@@ -57,5 +65,8 @@ func (ml *MarginLevels) consume(event MarginLevelsEvent) error {
 		return errors.Wrap(err, "converting margin levels proto to database entity failed")
 	}
 
-	return errors.Wrap(ml.store.Upsert(record), "inserting margin levels to SQL store failed")
+	record.SyntheticTime = ml.vegaTime.Add(time.Duration(ml.seqNum) * time.Microsecond)
+	record.SeqNum = ml.seqNum
+
+  return errors.Wrap(ml.store.Add(record), "inserting margin levels to SQL store failed")
 }