+++ CHANGELOG.md
@@ -21,7 +21,8 @@
 ### 🐛 Fixes
 - [451](https://github.com/vegaprotocol/data-node/issues/451) - Correct conversion of pending validator status
 - [391](https://github.com/vegaprotocol/data-node/issues/391) - Fix `OracleSpecs GraphQL` query returns error and null when there is no data.
-
+- [281](https://github.com/vegaprotocol/data-node/issues/281) - Fix Estimate Margin calculates incorrectly for Limit Orders
+- [482](https://github.com/vegaprotocol/data-node/issues/482) - Fan out event broker should only call listen once on source broker
 
 ## 0.49.3